Reformat projectile_angle_convel
